requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
structlog>=21.1.0
prometheus-client>=0.11.0
opentelemetry-api>=1.7.1
opentelemetry-sdk>=1.7.1
opentelemetry-instrumentation-fastapi>=0.30b1
opentelemetry-exporter-otlp>=1.7.1
httpx>=0.24.0
opentelemetry-instrumentation-httpx>=0.30b1 